+++ board-back/src/main/java/com/solpooh/boardback/service/BoardService.java
@@ -9,20 +9,20 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 
 public interface BoardService {
-    ResponseEntity<? super GetBoardResponseDto> getBoardDetail(String category, Integer boardNumber);
-    ResponseEntity<? super GetFavoriteListResponseDto> getFavoriteList(Integer boardNumber);
-    ResponseEntity<? super GetCommentListResponseDto> getCommentList(Integer boardNumber, Pageable pageable);
+    ResponseEntity<? super GetBoardResponseDto> getBoardDetail(String category, Long boardNumber);
+    ResponseEntity<? super GetFavoriteListResponseDto> getFavoriteList(Long boardNumber);
+    ResponseEntity<? super GetCommentListResponseDto> getCommentList(Long boardNumber, Pageable pageable);
     ResponseEntity<? super GetLatestBoardListResponseDto> getLatestBoardList(String category, Pageable pageable);
     ResponseEntity<? super GetTop3BoardListResponseDto> getTop3BoardList();
     ResponseEntity<? super GetSearchBoardListResponseDto> getSearchBoardList(String searchWord, String preSearchWord, Pageable pageable);
     ResponseEntity<? super GetUserBoardListResponseDto> getUserBoardList(String email, Pageable pageable);
     ResponseEntity<? super PostBoardResponseDto> postBoard(PostBoardRequestDto dto, String email);
-    ResponseEntity<? super PostCommentResponseDto> postComment(PostCommentRequestDto dto, Integer boardNumber, String email);
+    ResponseEntity<? super PostCommentResponseDto> postComment(PostCommentRequestDto dto, Long boardNumber, String email);
 
-    ResponseEntity<? super PutFavoriteResponseDto> putFavorite(Integer boardNumber, String email);
-    ResponseEntity<? super PatchBoardResponseDto> patchBoard(PatchBoardRequestDto dto, Integer boardNumber, String email);
-    ResponseEntity<? super PatchCommentResponseDto> patchComment(PatchCommentRequestDto dto, Integer boardNumber, Integer commentNumber, String email);
-    ResponseEntity<? super IncreaseViewCountResponseDto> increaseViewCount(Integer boardNumber);
-    ResponseEntity<? super DeleteBoardResponseDto> deleteBoard(Integer boardNumber, String email);
-    ResponseEntity<? super DeleteCommentResponseDto> deleteComment(Integer boardNumber, Integer commentNumber, String email);
+    ResponseEntity<? super PutFavoriteResponseDto> putFavorite(Long boardNumber, String email);
+    ResponseEntity<? super PatchBoardResponseDto> patchBoard(PatchBoardRequestDto dto, Long boardNumber, String email);
+    ResponseEntity<? super PatchCommentResponseDto> patchComment(PatchCommentRequestDto dto, Long boardNumber, Long commentNumber, String email);
+    ResponseEntity<? super IncreaseViewCountResponseDto> increaseViewCount(Long boardNumber);
+    ResponseEntity<? super DeleteBoardResponseDto> deleteBoard(Long boardNumber, String email);
+    ResponseEntity<? super DeleteCommentResponseDto> deleteComment(Long boardNumber, Long commentNumber, String email);
 }
